Fixing add and list issues
